requirements.txt
gradio>=4.0.0
torch>=2.0.0
transformers>=4.30.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
pyyaml>=6.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
pillow>=10.0.0
requests>=2.31.0
sentence-transformers>=2.2.0
nltk>=3.8.0
fastapi>=0.100.0
uvicorn>=0.22.0
